Huge overhall to automate.tin:
Now automations happen every 5 seconds, and can intelligently cast if you are in combat or not. In addition, automations will try up to 12 times, will not be attempted in the meta shop or the temple, and will be added to a list as they fall.

Basically instead of using a bunch of strings which may change at any moment to figure out when something has fallen, and then adding it to be done at the earliest possibility, we only add things to be recast when the affect is removed from the msdp affects list. If it is added before we have attempted to recast it, it will simply be removed from the pending list.

All this to say that automation should have just gotten a ton smarter! Please report bugs which almost certainly exist.